Map filter issue fix and error handle
