+++ components/appNavBar.js
@@ -38,7 +38,7 @@ const AppNavbar = () => {
                 <title>TicketMaster POC's</title>
                 <link rel='icon' href='/favicon.ico' />
                 <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous" />
-                <script></script>
+                <link rel="stylesheet" type="text/css" href="//fonts.googleapis.com/css?family=Open+Sans" />
             </Head>
             <Navbar color="" dark expand="sm" className="navBarBg">
                 <Container>
