+++ package.json
@@ -21,12 +21,14 @@
     "graphql": "^14.5.8",
     "graphql-tag": "^2.10.1",
     "highcharts-react-official": "^2.2.2",
+    "i": "^0.3.6",
     "moment": "^2.24.0",
     "moment-timezone": "^0.5.27",
     "isomorphic-fetch": "^2.2.1",
     "next": "^9.1.1",
     "node-fetch": "^2.6.0",
     "node-sass": "^4.13.0",
+    "npm": "^6.13.0",
     "npm-overlapping-marker-spiderfier": "^1.0.3",
     "react": "^16.11.0",
     "react-apollo": "^2.5.4",
