+++ pages/Events/events.scss
@@ -1,0 +1,6 @@
+
+.no-data {
+    text-align: center;
+    padding: 20px;
+    color: #f00;
+}
