+++ pages/Events/index.js
@@ -6,7 +6,11 @@ import { Container, Row, Col, Button, InputGroup, InputGroupAddon, InputGroupTex
 import Select from 'react-select';
 
 import Layout from '../../components/layout';
-import MyMapComponent from '../../components/Map';
+import MyMapComponent from '../../components/Map/_index';
+
+import EventMap from '../../components/Map';
+
+import './events.scss';
 
 
 const Events = (props) => {
@@ -94,10 +98,8 @@ const Events = (props) => {
         getEvents(selectedOption.value);
     }, []);
 
-    console.log(selectedOption)
-
     return (
-        <Layout>
+        <Layout className={'events-container'}>
             <Container>
                 <Row>
                     <Col>
@@ -120,8 +122,8 @@ const Events = (props) => {
                     </Col>
                 </Row>
             </Container>
-            {
-                markers.length &&   (
+            {/* {
+                markers.length && (
                     <MyMapComponent
                         markers={markers}
                         googleMapURL="https://maps.googleapis.com/maps/api/js?key=AIzaSyA6ytGTT86EN1ih4mm10nWQ7uV36khvcCc"
@@ -130,7 +132,12 @@ const Events = (props) => {
                         mapElement={<div style={{ height: `100%` }} />}
                     />
                 )
-            }
+            } */}
+            {markers.length ? (
+                <EventMap markers={markers} />
+            ) : (
+                    <h4 className = {'no-data'}>No data found!</h4>
+                )}
         </Layout>
     )
 }
