+++ pages/index.scss
@@ -1,3 +1,7 @@
+
+* {
+  font-family: "Open Sans"
+}
 .navBarBg{
     background-color: #026cdf;
 }
